pyproject.toml
[tool.poetry]
name = "course-work-7"
version = "0.1.0"
description = ""
authors = ["GAFFAROV ILKHOM"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
pillow = "^10.3.0"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
black = "^24.4.2"
isort = "^5.13.2"
ipython = "^8.25.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
